requirements.txt
Django==4.0.3
djangorestframework==3.13.1
psycopg2-binary==2.9.3
dj_database_url==0.5.0
pandas==1.4.2
gunicorn==20.0.4
whitenoise==4.1.4
# django-allauth==0.43.0
django-rest-auth==0.3.3
dj-rest-auth[with_social]==2.2.4
django-cors-headers==3.12.0